Use "When In Use" authorization for location services
StealthAssist does not need the "Always" authorization for location services. As of iOS 9, requesting the "When In Use" authorization in combination with background updates (location updates background mode plus `allowsBackgroundLocationUpdates` on CLLocationManager) is sufficient.

This change causes a blue status bar to appear when StealthAssist is running in the background and is using the device's location.