Fix bug preventing certain display colors from being selected
As of iOS 8 or 9, comparing two UIColor instances with -[isEqual:] no longer reliably works. This is likely due to an internal change that causes a slight round off error or other floating point loss of precision, so that when UIColor instances are serialized (as StealthAssist used to do when storing to NSUserDefaults) and then read back into memory later on, the deserialized color is not identical.

This change simply switches everything to be based off of the index of each color in the master array of colors.